+++ core-java/src/main/java/uk/jtoye/core/CoreApplication.java
@@ -3,12 +3,14 @@ package uk.jtoye.core;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.view.RedirectView;
 
 @SpringBootApplication
 @EnableAsync
+@EnableScheduling
 @RestController
 public class CoreApplication {